+++ tests/test_cli.py
@@ -102,7 +102,7 @@ def test_cli_shapesWGS84():
     assert result.exit_code == 0
     assert (
         result.output
-        == '{"bbox": [37.264008, -56.347191, 37.615571, -56.151464], "geometry": {"coordinates": [[[37.264008, -56.347191], [37.264008, -56.151464], [37.615571, -56.151464], [37.615571, -56.347191], [37.264008, -56.347191]]], "type": "Polygon"}, "id": "(106, 193, 9)", "properties": {"grid_crs": "EPSG:3395", "grid_name": "WorldMercatorWGS84Quad", "title": "XYZ tile (106, 193, 9)"}, "type": "Feature"}\n'
+        == '{"bbox": [-105.46875, 40.099155, -104.765625, 40.636956], "geometry": {"coordinates": [[[-105.46875, 40.099155], [-105.46875, 40.636956], [-104.765625, 40.636956], [-104.765625, 40.099155], [-105.46875, 40.099155]]], "type": "Polygon"}, "id": "(106, 193, 9)", "properties": {"grid_crs": "EPSG:3395", "grid_name": "WorldMercatorWGS84Quad", "title": "XYZ tile (106, 193, 9)"}, "type": "Feature"}\n'
     )